zram-introduce-zram-memory-tracking-fix
fix i386 printk warning

drivers/block/zram/zram_drv.c: In function read_block_state:
drivers/block/zram/zram_drv.c:683:9: warning: format %lu expects argument of type long unsigned int, but argument 4 has type ssize_t {aka int} [-Wformat=]

Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Randy Dunlap <rdunlap@infradead.org>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>

+++ drivers/block/zram/zram_drv.c
@@ -681,7 +681,8 @@ static ssize_t read_block_state(struct file *file, char __user *buf,
 		usec_rem = do_div(last_access, USEC_PER_SEC);
 		copied = snprintf(kbuf + written, count,
 			"%12lu %5lu.%06lu %c%c%c\n",
-			index, (unsigned long)last_access, usec_rem,
+			(unsigned long)index, (unsigned long)last_access,
+			usec_rem,
 			zram_test_flag(zram, index, ZRAM_SAME) ? 's' : '.',
 			zram_test_flag(zram, index, ZRAM_WB) ? 'w' : '.',
 			zram_test_flag(zram, index, ZRAM_HUGE) ? 'h' : '.');
